refactor: 将主键类型从 Long 更改为 Integer，优化数据模型和数据库访问层

+++ avalon-application/avalon-admin-server/src/main/java/io/github/lishangbu/avalon/admin/controller/dataset/BerryFlavorController.java
@@ -70,7 +70,7 @@ public class BerryFlavorController {
    * @param id 树果风味ID
    */
   @DeleteMapping("/{id:\\d+}")
-  public void deleteById(@PathVariable Long id) {
+  public void deleteById(@PathVariable Integer id) {
     berryFlavorService.removeById(id);
   }
 }
