+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/MoveAilmentRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface MoveAilmentRepository
-    extends ListCrudRepository<MoveAilment, Long>,
-        ListPagingAndSortingRepository<MoveAilment, Long> {}
+    extends ListCrudRepository<MoveAilment, Integer>,
+        ListPagingAndSortingRepository<MoveAilment, Integer> {}
