+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/MoveLearnMethodRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface MoveLearnMethodRepository
-    extends ListCrudRepository<MoveLearnMethod, Long>,
-        ListPagingAndSortingRepository<MoveLearnMethod, Long> {}
+    extends ListCrudRepository<MoveLearnMethod, Integer>,
+        ListPagingAndSortingRepository<MoveLearnMethod, Integer> {}
